// taurus-protect-sdk-typescript/src/internal/openapi/apis/ActionsApi.ts
export interface ActionServiceGetActionRequest {
    id: string;
}

export interface ActionServiceGetActionsRequest {
    limit?: string;
    offset?: string;
    ids?: Array<string>;
}

/**
 * ActionsApi - interface
 * 
 * @export
 * @interface ActionsApiInterface
 */
export interface ActionsApiInterface {
    /**
     * This endpoint returns a single action.  Required role: **Admin** or **AdminReadOnly**.
     * @summary Get an action
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ActionsApiInterface
     */
    actionServiceGetActionRaw(requestParameters: ActionServiceGetActionRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<TgvalidatordGetActionReply>>;

    /**
     * This endpoint returns a single action.  Required role: **Admin** or **AdminReadOnly**.
     * Get an action
     */
    actionServiceGetAction(requestParameters: ActionServiceGetActionRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<TgvalidatordGetActionReply>;

    /**
     * This endpoint lists all actions.  Required role: **Admin** or **AdminReadOnly**.
     * @summary List actions
     * @param {string} [limit] 
     * @param {string} [offset] 
     * @param {Array<string>} [ids] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ActionsApiInterface
     */
    actionServiceGetActionsRaw(requestParameters: ActionServiceGetActionsRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<TgvalidatordGetActionsReply>>;

    /**
     * This endpoint lists all actions.  Required role: **Admin** or **AdminReadOnly**.
     * List actions
     */
    actionServiceGetActions(requestParameters: ActionServiceGetActionsRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<TgvalidatordGetActionsReply>;

}

/**
 * 
 */
export class ActionsApi extends runtime.BaseAPI implements ActionsApiInterface {

    /**
     * This endpoint returns a single action.  Required role: **Admin** or **AdminReadOnly**.
     * Get an action
     */
    async actionServiceGetActionRaw(requestParameters: ActionServiceGetActionRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<TgvalidatordGetActionReply>> {
        if (requestParameters['id'] == null) {
            throw new runtime.RequiredError(
                'id',
                'Required parameter "id" was null or undefined when calling actionServiceGetAction().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        if (this.configuration && this.configuration.apiKey) {
            headerParameters["Authorization"] = await this.configuration.apiKey("Authorization"); // ApiKeyTPV1 authentication
        }

        const response = await this.request({
            path: `/api/rest/v1/actions/{id}`.replace(`{${"id"}}`, encodeURIComponent(String(requestParameters['id']))),
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        }, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => TgvalidatordGetActionReplyFromJSON(jsonValue));
    }

    /**
     * This endpoint returns a single action.  Required role: **Admin** or **AdminReadOnly**.
     * Get an action
     */
    async actionServiceGetAction(requestParameters: ActionServiceGetActionRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<TgvalidatordGetActionReply> {
        const response = await this.actionServiceGetActionRaw(requestParameters, initOverrides);
        return await response.value();
    }

    /**
     * This endpoint lists all actions.  Required role: **Admin** or **AdminReadOnly**.
     * List actions
     */
    async actionServiceGetActionsRaw(requestParameters: ActionServiceGetActionsRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<TgvalidatordGetActionsReply>> {
        const queryParameters: any = {};

        if (requestParameters['limit'] != null) {
            queryParameters['limit'] = requestParameters['limit'];
        }

        if (requestParameters['offset'] != null) {
            queryParameters['offset'] = requestParameters['offset'];
        }

        if (requestParameters['ids'] != null) {
            queryParameters['ids'] = requestParameters['ids'];
        }

        const headerParameters: runtime.HTTPHeaders = {};

        if (this.configuration && this.configuration.apiKey) {
            headerParameters["Authorization"] = await this.configuration.apiKey("Authorization"); // ApiKeyTPV1 authentication
        }

        const response = await this.request({
            path: `/api/rest/v1/actions`,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        }, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => TgvalidatordGetActionsReplyFromJSON(jsonValue));
    }

    /**
     * This endpoint lists all actions.  Required role: **Admin** or **AdminReadOnly**.
     * List actions
     */
    async actionServiceGetActions(requestParameters: ActionServiceGetActionsRequest = {}, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<TgvalidatordGetActionsReply> {
        const response = await this.actionServiceGetActionsRaw(requestParameters, initOverrides);
        return await response.value();
    }

}